03_webhook-api/requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
defusedxml==0.7.1
dill==0.3.7
filelock==3.12.4
Flask==2.3.3
frozenlist==1.4.0
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jira==3.5.2
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
packaging==23.1
pandas==2.1.3
Pillow==10.1.0
preshed==3.0.9
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.5
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rouge-score==0.1.2
sacremoses==0.0.53
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
translation-quality-estimator==0.1.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
weasel==0.3.4
Werkzeug==2.3.7
xxhash==3.4.1
yarl==1.9.2
